+++ samples/app/src/main/kotlin/com/bumble/appyx/app/node/samples/SamplesContainerNode.kt
@@ -2,26 +2,24 @@ package com.bumble.appyx.app.node.samples
 
 import android.os.Parcelable
 import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
-import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.ExperimentalUnitApi
 import androidx.compose.ui.unit.dp
-import com.bumble.appyx.app.node.helper.screenNode
-import com.bumble.appyx.app.node.onboarding.OnboardingContainerNode
+import com.bumble.appyx.app.node.cards.CardsExampleNode
+import com.bumble.appyx.app.node.slideshow.WhatsAppyxSlideShow
 import com.bumble.appyx.core.composable.Children
 import com.bumble.appyx.core.integrationpoint.LocalIntegrationPoint
 import com.bumble.appyx.core.modality.BuildContext
@@ -33,7 +31,6 @@ import com.bumble.appyx.navmodel.backstack.activeElement
 import com.bumble.appyx.navmodel.backstack.operation.newRoot
 import com.bumble.appyx.navmodel.backstack.operation.pop
 import com.bumble.appyx.navmodel.backstack.operation.push
-import com.bumble.appyx.navmodel.backstack.operation.replace
 import com.bumble.appyx.navmodel.backstack.transitionhandler.rememberBackstackSlider
 import com.bumble.appyx.sample.navigtion.compose.ComposeNavigationRoot
 import kotlinx.parcelize.Parcelize
@@ -76,9 +73,24 @@ class SamplesContainerNode(
     @ExperimentalComposeUiApi
     override fun resolve(navTarget: NavTarget, buildContext: BuildContext): Node =
         when (navTarget) {
-            NavTarget.SamplesListScreen -> screenNode(buildContext) { SamplesSelector(backStack) }
-            NavTarget.OnboardingScreen -> OnboardingContainerNode(buildContext)
-            NavTarget.ComposeNavigationScreen -> {
+            is NavTarget.SamplesListScreen -> SamplesSelectorNode(buildContext) { output ->
+                backStack.push(
+                    when (output) {
+                        is SamplesSelectorNode.Output.OpenCardsExample -> {
+                            NavTarget.CardsExample
+                        }
+                        is SamplesSelectorNode.Output.OpenComposeNavigation -> {
+                            NavTarget.ComposeNavigationScreen
+                        }
+                        is SamplesSelectorNode.Output.OpenOnboarding -> {
+                            NavTarget.OnboardingScreen
+                        }
+                    }
+                )
+            }
+            is NavTarget.CardsExample -> CardsExampleNode(buildContext)
+            is NavTarget.OnboardingScreen -> WhatsAppyxSlideShow(buildContext, isInPreviewMode = false)
+            is NavTarget.ComposeNavigationScreen -> {
                 node(buildContext) {
                     // compose-navigation fetches the integration point via LocalIntegrationPoint
                     CompositionLocalProvider(
@@ -88,7 +100,6 @@ class SamplesContainerNode(
                     }
                 }
             }
-            NavTarget.CardsExample -> CardsExampleNode(buildContext)
         }
 
     @ExperimentalUnitApi
@@ -96,7 +107,7 @@ class SamplesContainerNode(
     @ExperimentalComposeUiApi
     override fun onChildFinished(child: Node) {
         when (child) {
-            is OnboardingContainerNode -> backStack.newRoot(NavTarget.SamplesListScreen)
+            is WhatsAppyxSlideShow -> backStack.newRoot(NavTarget.SamplesListScreen)
             else -> super.onChildFinished(child)
         }
     }
@@ -105,41 +116,27 @@ class SamplesContainerNode(
     override fun View(modifier: Modifier) {
         val elementsState by backStack.elements.collectAsState()
 
-        if (elementsState.activeElement?.showBackButton == true) {
-            IconButton(onClick = { backStack.pop() }) {
-                Icon(
-                    imageVector = Icons.Filled.ArrowBack,
-                    contentDescription = "Back"
+        Box(Modifier.fillMaxSize()) {
+            if (elementsState.activeElement?.showBackButton == true) {
+                IconButton(onClick = { backStack.pop() }) {
+                    Icon(
+                        imageVector = Icons.Filled.ArrowBack,
+                        contentDescription = "Back"
+                    )
+                }
+            }
+            Column(
+                modifier = modifier
+                    .fillMaxSize()
+                    .padding(top = 40.dp)
+            ) {
+                Children(
+                    modifier = Modifier.fillMaxSize(),
+                    transitionHandler = rememberBackstackSlider(),
+                    navModel = backStack
                 )
-            }
-        }
-        Column(
-            modifier = modifier
-                .fillMaxSize()
-                .padding(top = 40.dp)
-        ) {
-            Children(
-                modifier = Modifier.fillMaxSize(),
-                transitionHandler = rememberBackstackSlider(),
-                navModel = backStack
-            )
-        }
-    }
-
-    @Composable
-    private fun SamplesSelector(backStack: BackStack<NavTarget>) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Button(onClick = { backStack.replace(NavTarget.OnboardingScreen) }) {
-                Text("What is Appyx?")
-            }
-            Button(onClick = { backStack.push(NavTarget.CardsExample) }) {
-                Text("Dating cards NavModel")
-            }
-            Button(onClick = { backStack.push(NavTarget.ComposeNavigationScreen) }) {
-                Text("Compose Navigation")
             }
         }
     }
 }
+